Handle haploid contigs in diploid file when reversing alleles
Desired behavior is: Haploid contigs like chrM are emitted to the unmapped file when converting from a personalized bedfile to reference bedfile using a diploid vci.

Current behavior: Haploid contig entries are emitted twice, once per _L/_R "haplotype".

Fix: Identify if a contig's suffix is not lr. If it is not, break the LR loop after emitting to the unmapped file, don't just continue to the right haplotype.